// nest-serve/src/chat/chat.d.ts
// 用户上传的文档
export interface uploadFileListType {
  fileName: string; // 文件名称
  fileSize: string; // 文件大小
  fileType: string; // 文件类型
  docId: string; // 文件id
}

// 用户和模型的对话结构
export interface MessagesType {
  role: 'user' | 'assistant' | 'system'; //角色
  content: string; //用户提问或者模型回复的内容
  uploadFileList?: {
    //携带的文档列表
    fileName: string; // 文件名称
    fileSize: string; // 文件大小
    fileType: string; // 文件类型
    docId: string; // 文件id
  }[];
  readFileData?: {
    type: 'readDocument' | 'queryKB'; //阅读文档 | 检索知识库
    statusInfo: 'inProgress' | 'completed'; //进行中 | 完毕
    promptInfo: string; //服务器返回的提示
    fileList: string[]; //处理的文件列表
  };
  displayContent?: string; //用户原始问题
  // 携带的图片
  uploadImageList?: {
    imagePath: string;
    mimetype: string;
    imageUrl: string;
  };
}